% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mirror.R
\name{hde}
\alias{hde}
\title{A dataset used for Homework 0.}
\description{
This is a part of the South African heart disease data, 
   we use classification tree method to provide a model for 
   determining whether a patient has heart disease (CHD).
}
