% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mirror.R
\name{MDS}
\alias{MDS}
\title{Multiple Data Splitting}
\usage{
MDS(y, X, q, m = 50)
}
\arguments{
\item{y}{vector of the response variable}

\item{X}{the design matrix}

\item{q}{the designated FDR level to be controlled at}

\item{m}{the number of independent random sample splits.}
}
\value{
a list with the selected features and the inclusion rate of all features.
}
\description{
Use Multiple Data Splitting to select features while controlling FDR for linear regression model.
}
\examples{
\dontrun{
n <- 100; p <- 5
beta <- c(0,1,1,2,0)
X <- matrix(rnorm(n*p),n,p)
y <- X \%*\% beta + rnorm(n)
MDS(y,X,0.1)$feature
}
}
