\name{SA24204176-package}
\alias{SA24204176-package}
\docType{package}
\title{
  Three methods about mirror to cotrol FDR for linear regression models
}
\description{
  Simultaneously finding multiple influential variables and 
    controlling the false discovery rate (FDR) for linear regression models 
    is a fundamental problem. This package provides three methods about 
    mirror (Gaussian mirror, data splitting and its multiple version) to 
    select features and control FDR at any nominal level.
}
\author{
Jun Zhou
Maintainer: Jun Zhou <zhoujun2020@mail.ustc.edu.cn>
}
\keyword{ package }
