% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Selection}
\alias{Selection}
\title{Selection features}
\usage{
Selection(M, q)
}
\arguments{
\item{M}{mirror statistic of all features}

\item{q}{the nominal FDR level}
}
\value{
the selected features that are considered significant
}
\description{
function to perform selection giving Mirror statistic and FDR level
}
